cli: Add pkg-config integration
The CLI currently requires users to manually specify library
paths and include directories. This makes it cumbersome to use
with libraries that provide pkg-config files, as users need to
look up these paths themselves.

This commit adds a --pkgconfig option that automatically
discovers library and include paths using pkg-config. When used
with --headers, it provides an easy way to generate
bindings:

  pybindlib --pkgconfig freerdp3 \
            --headers /usr/include/freerdp3/freerdp/freerdp.h